requirements/requirements-train.txt
dm-haiku==0.0.10
huggingface-hub
hydra-core==1.3
neptune-client==0.16.15
optax>=0.1.4
rlax>=0.1.4
tensorboardX==2.5.1
tqdm>=4.64.1
